Update Go install instructions
Using a modern version of Go (1.15.2 in my testing), the instructions are simplified to the point where I would not call it "harder". `GOPATH` and `GOBIN` are no longer required, and `GO111MODULE` is `auto` by default. Further, you don't need to separately install bitcomplete, and the installation steps aren't different for Windows anymore.

This patch fixes up the installation instructions to represent a modern Go environment.